requirements.txt
pyyaml>=6.0
numpy>=1.24
matplotlib>=3.7
tqdm>=4.65
pandas>=2.0
pytest>=7.0
